requirements.txt
# === Flask Backend ===
Flask==2.3.3
flask-cors==6.0.1
gunicorn==21.2.0

# === Web Scraping ===
playwright==1.44.0
beautifulsoup4==4.13.4
requests==2.31.0
aiohttp==3.9.5

# === Email / Website Parsing ===
lxml==5.2.1
html5lib==1.1
soupsieve==2.7

# === ChatGPT Message Generation ===
openai==1.33.0

# === CSV, Export, and Google Sheets ===
gspread==6.2.1
oauth2client==4.1.3

# === Misc Utilities ===
python-dotenv==1.1.1
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
MarkupSafe==3.0.2

# === Optional (if using FastAPI / Starlette instead of Flask) ===
# fastapi==0.115.14
# starlette==0.46.2
# uvicorn==0.35.0
